Switch to use 7zzs bundled with newer PM

+++ ports/xargon/Xargon 3 - Xargon's Fury.sh
@@ -23,17 +23,22 @@ GAMEDIR="/$directory/ports/xargon"
 export LD_LIBRARY_PATH="$GAMEDIR/libs.${DEVICE_ARCH}:$LD_LIBRARY_PATH"
 export SDL_GAMECONTROLLERCONFIG="$sdl_controllerconfig"
 
+if [ ! -f "$controlfolder/7zzs.${DEVICE_ARCH}" ]; then
+    pm_message "this port requires the latest portmaster to run, please go to https://portmaster.games/ for more info."
+    sleep 5
+    exit 1
+fi
+
 cd $GAMEDIR
 
 $ESUDO cp "$GAMEDIR/xrfile03.${DEVICE_ARCH}" "$GAMEDIR/xrfile03"
 
 $ESUDO chmod +x "$GAMEDIR/xrfile03"
-$ESUDO chmod +x "$GAMEDIR/7zzs.${DEVICE_ARCH}"
 $ESUDO chmod ugo+rw "$GAMEDIR/config.xr3"
 
 # Extract game files on 1st run
 if [ ! -f "$GAMEDIR/tiles.xr1" ]; then
-    "$GAMEDIR/7zzs.${DEVICE_ARCH}" x "$GAMEDIR/gamedata.7z" -o"$GAMEDIR/"
+    "$controlfolder/7zzs.${DEVICE_ARCH}" x "$GAMEDIR/gamedata.7z" -o"$GAMEDIR/"
 fi
 
 $GPTOKEYB "xrfile03" -c "$GAMEDIR/xargon.gptk" &
